% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Bumphunter_results.R
\docType{data}
\name{bumphunterRes_df}
\alias{bumphunterRes_df}
\title{Table of Bumphunter Results}
\format{A data frame containing 17 model fit measurements over 600 design
   points. The columns are described in the help documentation of
   \code{\link{ProcessBumphunterResults}}.}
\source{
Compiled via executing the \code{\link{WriteBumphunterResults}} and
   \code{\link{ProcessBumphunterResults}} functions.
}
\usage{
bumphunterRes_df
}
\description{
Results returned by the \code{\link{ProcessBumphunterResults}}
   function for the Bumphunter method.
}
\keyword{internal}
