% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pretty_means_sds.R
\name{CalcMeanSD}
\alias{CalcMeanSD}
\title{Calculate Mean and Standard Deviation}
\usage{
CalcMeanSD(vec, sigFigsMean = 2, sigFigsSD = 2, charact = TRUE,
  na.rm = TRUE)
}
\arguments{
\item{vec}{A numeric vector to summarize}

\item{sigFigsMean}{How many significant figures to display the mean? Defaults
to 2.}

\item{sigFigsSD}{How many significant figures to display the standard
deviations? Defaults to 2.}

\item{charact}{Should the results be returned as a character string in
"\code{xx.xx (x.xx)}" format? Defaults to \code{TRUE}. If \code{FALSE},
this function will return a data frame with column names \code{mean} and
\code{std_dev}.}

\item{na.rm}{Should missing values be removed? Defaults to \code{TRUE}.}
}
\value{
A character string in the form of "mean (sd)", if \code{charact} is
   set to \code{TRUE}. Otherwise, a data frame of the mean and standard
   deviation of \code{vec}.
}
\description{
Given a vector, calculate its mean and standard deviation, then
   display the results in a "pretty" manner.
}
\details{
This function is for help in building large tables of output wherein
   the user may need many entries in the form "mean (std dev)".
}
\examples{
   # Called by the script in inst/docs/Method_compare_report_20180705.Rmd
   CalcMeanSD(1:10)
}
\keyword{internal}
