% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cpg_list_to_table.R
\name{ConvertCPGList}
\alias{ConvertCPGList}
\title{CPG List-to-Table Conversion}
\usage{
ConvertCPGList(cpgs_ls, methylval_df)
}
\arguments{
\item{cpgs_ls}{a list where each item has CPG IDs in a cluster}

\item{methylval_df}{a dataframe of beta values}
}
\value{
a data table with
   \itemize{
     \item{\code{rows}}{cluster 1,1,1, 2,2,2}
     \item{\code{columns}}{sample IDs}
     \item{\code{value}}{mvalue or beta values}
   }
}
\description{
Convert a list of CPGs to a table
}
\examples{
  # Called by the script in inst/old_scripts/1_Aclust_data_import.R
}
\keyword{internal}
