% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_DMRcate_results.R
\docType{data}
\name{dmrcateRes_df}
\alias{dmrcateRes_df}
\title{Table of DMRcate Results}
\format{A data frame containing 17 model fit measurements over 1000 design
   points. The columns are described in \code{\link{ProcessDMRcateResults}}'s
   help documentation.}
\source{
Compiled via executing the \code{\link{WriteDMRcateResults}} and
   \code{\link{ProcessDMRcateResults}} functions.
}
\usage{
dmrcateRes_df
}
\description{
Results returned by the \code{\link{ProcessDMRcateResults}}
   function for the DMRcate method.
}
\keyword{internal}
