% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ProbeLasso_results.R
\docType{data}
\name{probeLassoRes_df}
\alias{probeLassoRes_df}
\title{Table of ProbeLasso Results}
\format{A data frame containing 17 model fit measurements over 2400 design
   points. The columns are described in the help documentation of
   \code{\link{ProcessProbeLassoResults}}.}
\source{
Compiled via executing the \code{\link{WriteProbeLassoResults}} and
   \code{\link{ProcessProbeLassoResults}} functions.
}
\usage{
probeLassoRes_df
}
\description{
Results returned by the \code{\link{ProcessProbeLassoResults}}
   function for the ProbeLasso method.
}
\keyword{internal}
